refactor(timestamp): extract video navigation logic into shared utility
Consolidate duplicate timestamp navigation code into a single
navigateVideoToTimestamp function in dom.ts. This function handles
backward compatibility for both chat replay timestamps (milliseconds)
and legacy cached timestamp links (seconds).

Changes:
- Add navigateVideoToTimestamp to dom.ts with JSDoc
- Update commentInteractions.ts to use shared function
- Update appController.ts to use shared function
- Export new function from assist.ts
