+++ app/src/source/utils/assist.ts
@@ -53,7 +53,8 @@ export {
     getPiP,
     initShowBarFAQ,
     initShowViewMode,
-    getRandomComment
+    getRandomComment,
+    navigateVideoToTimestamp
 } from './dom';
 
 export {
