+++ app/src/source/web-resources/ui/commentInteractions.ts
@@ -1,4 +1,4 @@
-import { removeNodeList } from '../../utils/dom';
+import { removeNodeList, navigateVideoToTimestamp } from '../../utils/dom';
 import { iconCollapse, iconExpand } from '../../utils/icons';
 import { ICommentsFuseResult } from '../../utils/interfaces/i_types';
 import { renderComment } from '../../utils/renderView';
@@ -197,20 +197,7 @@ function handleGotoChatVideo(event: Event): void {
     const video = document.getElementsByTagName('video')[0];
     if (!video) return;
 
-    const timeValue = target.dataset.offsetvideo;
-    if (!timeValue) return;
-
-    const parsed = Number.parseInt(timeValue, 10);
-    if (!Number.isFinite(parsed)) return;
-
-    // Backward compatibility logic:
-    // - ycs_goto_chat: Chat "Go to" button (milliseconds)
-    // - Others: Old cached timestamp links (seconds)
-    if (target.classList.contains('ycs_goto_chat')) {
-        video.currentTime = parsed / 1000;
-    } else {
-        video.currentTime = parsed;
-    }
+    navigateVideoToTimestamp(target, video);
 }
 
 function handleGotoCommentTime(event: Event): void {
@@ -222,13 +209,7 @@ function handleGotoCommentTime(event: Event): void {
     const video = document.getElementsByTagName('video')[0];
     if (!video) return;
 
-    const seconds = target.dataset.offsetvideo;
-    if (!seconds) return;
-
-    const secondsValue = Number.parseInt(seconds, 10);
-    if (Number.isFinite(secondsValue)) {
-        video.currentTime = secondsValue;
-    }
+    navigateVideoToTimestamp(target, video);
 }
 
 function collectRepliesForComment(
